Disallow TestCases to add/remove from InternalChildren
